streamlit_app/requirements.txt
streamlit 
psycopg2-binary
langchain
python-dotenv
tiktoken
